// admin-deploy-temp/types/chat.ts
export interface UserChat {
  id: string;
  participantId: string;
  participantName: string;
  participantImage: string;
  lastMessageAt: string | null;
  lastMessage?: string;
  unreadCount?: number;
  createdAt: string;
}